bot/pyproject.toml
[project]
name = "bot"
version = "0.2.3"
description = "Nora Bot"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "aiogram>=3.19.0",
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "pydantic-settings>=2.8.1",
]
